+++ linkerd/app/core/src/metrics.rs
@@ -96,14 +96,6 @@ pub struct RouteAuthzLabels {
     pub authz: Arc<policy::Meta>,
 }
 
-/// Labels referencing an inbound HTTP local rate limit.
-#[derive(Clone, Debug, Eq, PartialEq, Hash)]
-pub struct HTTPLocalRateLimitLabels {
-    pub server: ServerLabel,
-    pub rate_limit: Option<Arc<policy::Meta>>,
-    pub error: String,
-}
-
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
 pub struct OutboundEndpointLabels {
     pub server_id: tls::ConditionalClientTls,
@@ -388,24 +380,6 @@ impl FmtLabels for RouteAuthzLabels {
             self.authz.kind(),
             self.authz.name(),
         )
-    }
-}
-
-impl FmtLabels for HTTPLocalRateLimitLabels {
-    fn fmt_labels(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.server.fmt_labels(f)?;
-        if let Some(rl) = &self.rate_limit {
-            write!(
-                f,
-                ",ratelimit_group=\"{}\",ratelimit_kind=\"{}\",ratelimit_name=\"{}\",ratelimit_error=\"{}\"",
-                rl.group(),
-                rl.kind(),
-                rl.name(),
-                self.error,
-            )
-        } else {
-            write!(f, ",ratelimit_error=\"{}\"", self.error)
-        }
     }
 }
